requirements.txt
pandas
typing
ast
os
pandasql
sqlalchemy
csv
openai
time
tiktoken
recognizers_suite
recognizers-text-suite
unicodedata
re
fuzzywuzzy
argparse
json
datasets
ragas
evaluate
nltk
absl
sqlite3
google.generativeai
